+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "astraea",
-  "version": "0.2.3",
+  "version": "0.3.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
