+++ src/app/(typef)/games/palworld/servers/[serverId]/(local)/commands/broadcast/broadcast.action.ts
@@ -4,9 +4,7 @@ import { action } from "@/(global)/lib/request/next-safe-action";
 import { z } from "zod";
 
 import { gamelist } from "@/(global)/meta/gamedata";
-import { actionWithErrorHandling } from "@/(global)/lib/request/next-safe-action";
-import { announceMessage } from "@/(global)/lib/palworld/rest-api";
-import { getServerAddress } from "@/(global)/lib/cloud-provider/server";
+import { broadcast } from "./broadcast.service";
 
 const broadcastSchema = z.object({
     game: z.enum(gamelist),
@@ -17,8 +15,6 @@ const broadcastSchema = z.object({
 export const broadcastAction = action(
     broadcastSchema,
     async ({ game, serverId, message }) => {
-        const serverAddress = await getServerAddress(game, serverId);
-
-        await announceMessage(serverAddress, message);
+        broadcast({ game, serverId, message });
     }
 );
